Get rid of stdint.h

+++ include/compiler.h
@@ -3,7 +3,6 @@
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <stdint.h>
 #include <string.h>
 
 #include "info.h"
@@ -59,8 +58,8 @@ struct label_s {
 };
 
 int compiler_run(char*, char*);
-void compiler_parameters(char*, char*, uint32_t*, uint32_t*, int, struct label_s*, unsigned int);
-void compiler_store_parameter(char*, uint32_t*, char*, int, struct label_s*, unsigned int);
-void compiler_label(char*, char*, uint32_t*, uint32_t*, int, struct label_s*, unsigned int*);
+void compiler_parameters(char*, char*, unsigned long int*, unsigned long int*, int, struct label_s*, unsigned int);
+void compiler_store_parameter(char*, unsigned long int*, char*, int, struct label_s*, unsigned int);
+void compiler_label(char*, char*, unsigned long int*, unsigned long int*, int, struct label_s*, unsigned int*);
 
 #endif
